refactor auto replies adn templates
