+++ src/utils/api/index.js
@@ -1,4 +1,5 @@
 import Account from '@/utils/api/account.js';
+import AutoReply from '@/utils/api/auto_reply.js';
 import Blacklist from '@/utils/api/blacklist.js';
 import Conversations from '@/utils/api/conversations.js';
 import Devices from '@/utils/api/devices.js';
@@ -7,9 +8,11 @@ import Folders from '@/utils/api/folders.js';
 import Messages from '@/utils/api/messages.js';
 import ScheduledMessages from '@/utils/api/scheduled_messages.js';
 import Stream from '@/utils/api/stream.js';
+import Templates from '@/utils/api/templates.js';
 
 export {
     Account,
+    AutoReply,
     Blacklist,
     Conversations,
     Devices,
@@ -17,5 +20,6 @@ export {
     Folders,
     Messages,
     ScheduledMessages,
-    Stream
+    Stream,
+    Templates
 }
